Convert between period, semimajor axis, and total mass for Keplerian orbit.
For cases where a quick and easy conversion between period and semimajor axis
is needed for some sort of binary system, this function will be able to do the
conversion relatively quickly by taking one aspect to be fixed, and doing the
rest of the conversions through the Astropy Quantitity framework.

What needs to be added now is testing guidelines.